+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `sign_field` no longer changes field appearances
+
 - `style_text` now adapts Tr (text rendering mode) better to user parameters
 
 ### Fixed